requirements.txt
requests
pyyaml
termcolor 
pick